Fix multiplication overflow-to-larger-type CodeQL alerts in LLD hashmap/hashtree

+++ src/LLD/hashtree.cpp
@@ -189,7 +189,7 @@ namespace LLD
         if(!filesystem::exists(file))
         {
             /* Build a vector with empty bytes to flush to disk. */
-            std::vector<uint8_t> vSpace(HASHMAP_TOTAL_BUCKETS * HASHMAP_KEY_ALLOCATION, 0xff);
+            std::vector<uint8_t> vSpace(static_cast<std::size_t>(HASHMAP_TOTAL_BUCKETS) * HASHMAP_KEY_ALLOCATION, 0xff);
 
             /* Flush the empty keychain file to disk. */
             std::fstream stream(file, std::ios::out | std::ios::binary | std::ios::trunc);
@@ -358,7 +358,7 @@ namespace LLD
                 if(!filesystem::exists(filename))
                 {
                     /* Blank vector to write empty space in new disk file. */
-                    static const std::vector<uint8_t> vSpace(HASHMAP_TOTAL_BUCKETS * HASHMAP_KEY_ALLOCATION, 0xff);
+                    static const std::vector<uint8_t> vSpace(static_cast<std::size_t>(HASHMAP_TOTAL_BUCKETS) * HASHMAP_KEY_ALLOCATION, 0xff);
 
                     /* Write the blank data to the new file handle. */
                     std::fstream stream(filename, std::ios::out | std::ios::binary | std::ios::trunc);
